requirements.txt
paralleldots
tk
python-dotenv